# custom_gpt_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Digital Product Auto-Publisher Agent API",
    "description": "Autonomous Agent API to create and publish digital products to 20+ marketplaces & websites from Custom GPTs.",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/stores/add": {
      "post": {
        "summary": "Add New Store",
        "description": "Called by Custom GPT to dynamically connect and register ANY new website or marketplace.",
        "operationId": "add_new_store_api_stores_add_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddStoreRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddStoreResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/session/sync": {
      "post": {
        "summary": "Sync Device Session",
        "description": "Syncs authenticated session from any device (phone, laptop, PC) into the Cloud Vault.",
        "operationId": "sync_device_session_api_session_sync_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionSyncRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/publish": {
      "post": {
        "summary": "Publish Digital Product",
        "description": "Called by Custom GPT to build and publish a digital product to all platforms.",
        "operationId": "publish_digital_product_api_publish_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AddStoreRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Display name for the website or store (e.g. 'My WordPress Store')"
          },
          "url": {
            "type": "string",
            "title": "Url",
            "description": "Base URL of the website (e.g. 'https://mystore.com')"
          },
          "type": {
            "type": "string",
            "title": "Type",
            "description": "Type of store: 'woocommerce', 'shopify', 'custom_webhook', or 'marketplace'",
            "default": "woocommerce"
          },
          "api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key",
            "description": "Consumer Key or Shopify Access Token (if API based)"
          },
          "api_secret": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Secret",
            "description": "Consumer Secret for WooCommerce (optional)"
          },
          "new_product_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "New Product Url",
            "description": "Custom 1-click new product listing URL if marketplace (e.g. Etsy, etc.)"
          }
        },
        "type": "object",
        "required": [
          "name",
          "url"
        ],
        "title": "AddStoreRequest"
      },
      "AddStoreResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "store": {
            "type": "object",
            "title": "Store"
          }
        },
        "type": "object",
        "required": [
          "status",
          "message",
          "store"
        ],
        "title": "AddStoreResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LinkedInPromoItem": {
        "properties": {
          "headline": {
            "type": "string",
            "title": "Headline"
          },
          "post_copy": {
            "type": "string",
            "title": "Post Copy"
          },
          "one_click_share_url": {
            "type": "string",
            "title": "One Click Share Url"
          }
        },
        "type": "object",
        "required": [
          "headline",
          "post_copy",
          "one_click_share_url"
        ],
        "title": "LinkedInPromoItem"
      },
      "PlatformStatusItem": {
        "properties": {
          "platform": {
            "type": "string",
            "title": "Platform"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "platform",
          "status",
          "url",
          "message"
        ],
        "title": "PlatformStatusItem"
      },
      "PublishRequest": {
        "properties": {
          "prompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt",
            "description": "Idea/topic for the digital product to generate (e.g. 'Gym Fitness & Diet Tracker')"
          },
          "product_file": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product File",
            "description": "Existing file name in workspace if publishing an already created file"
          },
          "format_type": {
            "type": "string",
            "title": "Format Type",
            "description": "Type of product: 'spreadsheet', 'document', or 'code'",
            "default": "spreadsheet"
          },
          "no_api": {
            "type": "boolean",
            "title": "No Api",
            "description": "Whether to use Zero-API browser automation (default True)",
            "default": true
          },
          "auto_approve": {
            "type": "boolean",
            "title": "Auto Approve",
            "description": "Whether to auto-approve pre-publish checklist",
            "default": true
          },
          "custom_store_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Custom Store Url",
            "description": "Direct URL of user's personal WooCommerce or Shopify store"
          },
          "custom_store_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Custom Store Type",
            "description": "Type of store: 'woocommerce' or 'shopify'"
          },
          "custom_store_api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Custom Store Api Key",
            "description": "Shopify Access Token or WooCommerce Consumer Key"
          },
          "custom_store_api_secret": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Custom Store Api Secret",
            "description": "WooCommerce Consumer Secret (if WooCommerce)"
          }
        },
        "type": "object",
        "title": "PublishRequest"
      },
      "PublishResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "product_title": {
            "type": "string",
            "title": "Product Title"
          },
          "format_category": {
            "type": "string",
            "title": "Format Category"
          },
          "bundle_zip": {
            "type": "string",
            "title": "Bundle Zip"
          },
          "total_platforms": {
            "type": "integer",
            "title": "Total Platforms"
          },
          "platforms": {
            "items": {
              "$ref": "#/components/schemas/PlatformStatusItem"
            },
            "type": "array",
            "title": "Platforms"
          },
          "markdown_report": {
            "type": "string",
            "title": "Markdown Report"
          },
          "linkedin_promo": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LinkedInPromoItem"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "status",
          "product_title",
          "format_category",
          "bundle_zip",
          "total_platforms",
          "platforms",
          "markdown_report"
        ],
        "title": "PublishResponse"
      },
...
      "SessionSyncRequest": {
        "properties": {
          "platform": {
            "type": "string",
            "title": "Platform",
            "description": "Platform name (e.g. 'Gumroad', 'Etsy', 'Shopify')"
          },
          "auth_token_or_cookie": {
            "type": "string",
            "title": "Auth Token Or Cookie",
            "description": "Session token or cookie string"
          },
          "expires_in_days": {
            "type": "integer",
            "title": "Expires In Days",
            "description": "Session validity duration in days",
            "default": 180
          },
          "device_name": {
            "type": "string",
            "title": "Device Name",
            "description": "Name of device performing the sync",
            "default": "User Device"
          }
        },
        "type": "object",
        "required": [
          "platform",
          "auth_token_or_cookie"
        ],
        "title": "SessionSyncRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
  "servers": [
    {
      "url": "https://auto-publish-agents.vercel.app"
    }
  ]
}